Fixed bug #4, regarding ENU coordinate shifts. Problem was that the reference point wasn't staying constant, giving the appearance that the coordinate systems were erroneous. Removed all instances of 'Int64' and changed to 'Int' Confirmed LLA, ECEF, and ENU transformations with Matlab's Mapping toolbox.

+++ src/highways.jl
@@ -41,7 +41,7 @@ end
 ### Gather highway data from OSM element ###
 function getHighwayData( highway::LightXML.XMLElement, class::String="" )
     oneway = false
-    nodes = Int64[]
+    nodes = Int[]
     road_name = ""
 
     # Get way ID
